refactor(e2e): extract PPTX test into independent spec file
Move the pptx-generator skill test from full-electron.spec.ts into its
own pptx-generator.spec.ts so it can be run independently.

Keep helpers/verify-pptx.py as a shared helper.
